refactor: getAppArticles API 에 query params 추가 및 api 타입을 /api/types 로 분리
